requirements.txt
# Azure Functions for ZipCode Population Density Assessment
azure-functions>=1.18.0
azure-functions-worker>=1.1.9
pandas>=2.1.0
numpy>=1.24.0
geopy>=2.3.0
pydantic>=2.0.0
requests>=2.31.0